requirements.txt
numpy
scikit-learn
Django==2.2.3
django-countries==5.3.3
django-haystack==2.8.1
django-language-field==0.0.3
django-multiselectfield==0.1.8
djangorestframework==3.9.4
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
Whoosh==2.7.4
pandas==0.24.2
django-notifications-hq==1.5.0